fix in numeric test to work on windows

+++ tests/numeric_test.py
@@ -14,19 +14,19 @@ def test_container():
     assert isinstance(container.a, sarray)
     assert isinstance(container.a, sarray[int])
     assert not isinstance(container.a, sarray[float])
-    assert isinstance(container.a[0], np.int64)
+    assert isinstance(container.a[0], np.integer)
 
     assert isinstance(container.b, sarray)
     assert isinstance(container.b, sarray[float])
-    assert isinstance(container.b[0], np.float64)
+    assert isinstance(container.b[0], np.float)
 
     assert isinstance(container.c, carray)
     assert isinstance(container.c, carray[int])
-    assert isinstance(container.c[0], np.int64)
+    assert isinstance(container.c[0], np.integer)
 
     assert isinstance(container.d, barray)
     assert isinstance(container.d, barray[int])
-    assert isinstance(container.d[0], np.int64)
+    assert isinstance(container.d[0], np.integer)
     assert 'd' in record_fields(container)
 
     lst = ['a', 'b']
